refactor: move webapp exceptions out of rest package
