+++ operate/webapp/src/main/java/io/camunda/operate/webapp/elasticsearch/QueryHelper.java
@@ -24,9 +24,9 @@ import io.camunda.operate.property.OperateProperties;
 import io.camunda.operate.store.IncidentStore;
 import io.camunda.operate.util.CollectionUtil;
 import io.camunda.operate.util.ElasticsearchUtil;
+import io.camunda.operate.webapp.exception.InvalidRequestException;
 import io.camunda.operate.webapp.rest.dto.listview.ListViewQueryDto;
 import io.camunda.operate.webapp.rest.dto.listview.VariablesQueryDto;
-import io.camunda.operate.webapp.rest.exception.InvalidRequestException;
 import io.camunda.operate.webapp.security.permission.PermissionsService;
 import io.camunda.webapps.schema.descriptors.template.ListViewTemplate;
 import io.camunda.webapps.schema.entities.flownode.FlowNodeState;
