+++ operate/webapp/src/main/java/io/camunda/operate/webapp/elasticsearch/reader/VariableReader.java
@@ -17,8 +17,8 @@ import io.camunda.operate.conditions.ElasticsearchCondition;
 import io.camunda.operate.exceptions.OperateRuntimeException;
 import io.camunda.operate.property.OperateProperties;
 import io.camunda.operate.util.ElasticsearchUtil;
+import io.camunda.operate.webapp.exception.NotFoundException;
 import io.camunda.operate.webapp.rest.dto.VariableDto;
-import io.camunda.operate.webapp.rest.exception.NotFoundException;
 import io.camunda.webapps.schema.descriptors.template.VariableTemplate;
 import io.camunda.webapps.schema.entities.VariableEntity;
 import java.io.IOException;
