+++ operate/webapp/src/main/java/io/camunda/operate/webapp/exception/InternalAPIException.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.operate.webapp.rest.exception;
+package io.camunda.operate.webapp.exception;
 
 import java.util.UUID;
