+++ operate/webapp/src/main/java/io/camunda/operate/webapp/rest/dto/operation/MigrationPlanDto.java
@@ -7,8 +7,8 @@
  */
 package io.camunda.operate.webapp.rest.dto.operation;
 
+import io.camunda.operate.webapp.exception.InvalidRequestException;
 import io.camunda.operate.webapp.rest.dto.RequestValidator;
-import io.camunda.operate.webapp.rest.exception.InvalidRequestException;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import java.util.Objects;
